requirements.txt
crewai
crewai-tools
langchain
langchain-community
langchain-openai
pymilvus
pandas
flask
python-dotenv
kaggle
tqdm
faiss-cpu
beautifulsoup4
requests
markupsafe
tiktoken